POS: Refactor NfcDialog into PosPaymentDialog
